pyproject.toml
[tool.poetry]
name = "gendiff"
version = "0.1.0"
description = ""
authors = ["immuneQQ <vitalyasatka@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
test1 = "gendiff.scripts.test1:main"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
